Move ExternalAccounts to package extsvc
