+++ enterprise/cmd/frontend/db/perms_store_mock.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/authz"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 )
 
 type MockPerms struct {
@@ -12,6 +13,6 @@ type MockPerms struct {
 	LoadUserPermissions        func(ctx context.Context, p *authz.UserPermissions) error
 	LoadUserPendingPermissions func(ctx context.Context, p *authz.UserPendingPermissions) error
 	SetRepoPermissions         func(ctx context.Context, p *authz.RepoPermissions) error
-	SetRepoPendingPermissions  func(ctx context.Context, accounts *ExternalAccounts, p *authz.RepoPermissions) error
+	SetRepoPendingPermissions  func(ctx context.Context, accounts *extsvc.ExternalAccounts, p *authz.RepoPermissions) error
 	ListPendingUsers           func(ctx context.Context) ([]string, error)
 }
