+++ enterprise/cmd/frontend/db/perms_store_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/gitchander/permutation"
 	"github.com/keegancsmith/sqlf"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/authz"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"golang.org/x/sync/errgroup"
 )
 
@@ -543,7 +544,7 @@ func testPermsStore_LoadUserPendingPermissions(db *sql.DB) func(*testing.T) {
 			s := NewPermsStore(db, clock)
 			defer cleanupPermsTables(t, s)
 
-			accounts := &ExternalAccounts{
+			accounts := &extsvc.ExternalAccounts{
 				ServiceType: "sourcegraph",
 				ServiceID:   "https://sourcegraph.com/",
 				AccountIDs:  []string{"bob"},
@@ -574,7 +575,7 @@ func testPermsStore_LoadUserPendingPermissions(db *sql.DB) func(*testing.T) {
 			s := NewPermsStore(db, clock)
 			defer cleanupPermsTables(t, s)
 
-			accounts := &ExternalAccounts{
+			accounts := &extsvc.ExternalAccounts{
 				ServiceType: "sourcegraph",
 				ServiceID:   "https://sourcegraph.com/",
 				AccountIDs:  []string{"alice"},
@@ -605,7 +606,7 @@ func testPermsStore_LoadUserPendingPermissions(db *sql.DB) func(*testing.T) {
 			s := NewPermsStore(db, clock)
 			defer cleanupPermsTables(t, s)
 
-			accounts := &ExternalAccounts{
+			accounts := &extsvc.ExternalAccounts{
 				ServiceType: "sourcegraph",
 				ServiceID:   "https://sourcegraph.com/",
 				AccountIDs:  []string{"alice", "bob"},
@@ -769,7 +770,7 @@ func checkRepoPendingPermsTable(ctx context.Context, s *PermsStore, bindIDs map[
 
 func testPermsStore_SetRepoPendingPermissions(db *sql.DB) func(*testing.T) {
 	type update struct {
-		accounts *ExternalAccounts
+		accounts *extsvc.ExternalAccounts
 		perm     *authz.RepoPermissions
 	}
 	tests := []struct {
@@ -782,7 +783,7 @@ func testPermsStore_SetRepoPendingPermissions(db *sql.DB) func(*testing.T) {
 			name: "empty",
 			updates: []update{
 				{
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  nil,
@@ -798,7 +799,7 @@ func testPermsStore_SetRepoPendingPermissions(db *sql.DB) func(*testing.T) {
 			name: "add",
 			updates: []update{
 				{
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  []string{"alice"},
@@ -808,7 +809,7 @@ func testPermsStore_SetRepoPendingPermissions(db *sql.DB) func(*testing.T) {
 						Perm:   authz.Read,
 					},
 				}, {
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  []string{"alice", "bob"},
@@ -818,7 +819,7 @@ func testPermsStore_SetRepoPendingPermissions(db *sql.DB) func(*testing.T) {
 						Perm:   authz.Read,
 					},
 				}, {
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  []string{"cindy", "david"},
@@ -845,7 +846,7 @@ func testPermsStore_SetRepoPendingPermissions(db *sql.DB) func(*testing.T) {
 			name: "add and update",
 			updates: []update{
 				{
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  []string{"alice", "bob"},
@@ -855,7 +856,7 @@ func testPermsStore_SetRepoPendingPermissions(db *sql.DB) func(*testing.T) {
 						Perm:   authz.Read,
 					},
 				}, {
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  []string{"bob", "cindy"},
@@ -865,7 +866,7 @@ func testPermsStore_SetRepoPendingPermissions(db *sql.DB) func(*testing.T) {
 						Perm:   authz.Read,
 					},
 				}, {
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  []string{"alice", "bob"},
@@ -875,7 +876,7 @@ func testPermsStore_SetRepoPendingPermissions(db *sql.DB) func(*testing.T) {
 						Perm:   authz.Read,
 					},
 				}, {
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  []string{"cindy", "david"},
@@ -901,7 +902,7 @@ func testPermsStore_SetRepoPendingPermissions(db *sql.DB) func(*testing.T) {
 			name: "add and clear",
 			updates: []update{
 				{
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  []string{"alice", "bob", "cindy"},
@@ -911,7 +912,7 @@ func testPermsStore_SetRepoPendingPermissions(db *sql.DB) func(*testing.T) {
 						Perm:   authz.Read,
 					},
 				}, {
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  []string{},
@@ -980,7 +981,7 @@ func testPermsStore_SetRepoPendingPermissions(db *sql.DB) func(*testing.T) {
 
 func testPermsStore_ListPendingUsers(db *sql.DB) func(t *testing.T) {
 	type update struct {
-		accounts *ExternalAccounts
+		accounts *extsvc.ExternalAccounts
 		perm     *authz.RepoPermissions
 	}
 	tests := []struct {
@@ -996,7 +997,7 @@ func testPermsStore_ListPendingUsers(db *sql.DB) func(t *testing.T) {
 			name: "has user with pending permissions",
 			updates: []update{
 				{
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  []string{"alice"},
@@ -1013,7 +1014,7 @@ func testPermsStore_ListPendingUsers(db *sql.DB) func(t *testing.T) {
 			name: "has user but with empty object_ids",
 			updates: []update{
 				{
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  []string{"bob@example.com"},
@@ -1023,7 +1024,7 @@ func testPermsStore_ListPendingUsers(db *sql.DB) func(t *testing.T) {
 						Perm:   authz.Read,
 					},
 				}, {
-					accounts: &ExternalAccounts{
+					accounts: &extsvc.ExternalAccounts{
 						ServiceType: "sourcegraph",
 						ServiceID:   "https://sourcegraph.com/",
 						AccountIDs:  nil,
@@ -1071,7 +1072,7 @@ func testPermsStore_ListPendingUsers(db *sql.DB) func(t *testing.T) {
 
 func testPermsStore_GrantPendingPermissions(db *sql.DB) func(t *testing.T) {
 	type pending struct {
-		accounts *ExternalAccounts
+		accounts *extsvc.ExternalAccounts
 		perm     *authz.RepoPermissions
 	}
 	type update struct {
@@ -1123,7 +1124,7 @@ func testPermsStore_GrantPendingPermissions(db *sql.DB) func(t *testing.T) {
 					},
 					pendings: []pending{
 						{
-							accounts: &ExternalAccounts{
+							accounts: &extsvc.ExternalAccounts{
 								ServiceType: "sourcegraph",
 								ServiceID:   "https://sourcegraph.com/",
 								AccountIDs:  []string{"alice"},
@@ -1133,7 +1134,7 @@ func testPermsStore_GrantPendingPermissions(db *sql.DB) func(t *testing.T) {
 								Perm:   authz.Read,
 							},
 						}, {
-							accounts: &ExternalAccounts{
+							accounts: &extsvc.ExternalAccounts{
 								ServiceType: "sourcegraph",
 								ServiceID:   "https://sourcegraph.com/",
 								AccountIDs:  []string{"bob"},
@@ -1192,7 +1193,7 @@ func testPermsStore_GrantPendingPermissions(db *sql.DB) func(t *testing.T) {
 					},
 					pendings: []pending{
 						{
-							accounts: &ExternalAccounts{
+							accounts: &extsvc.ExternalAccounts{
 								ServiceType: "sourcegraph",
 								ServiceID:   "https://sourcegraph.com/",
 								AccountIDs:  []string{"alice"},
@@ -1202,7 +1203,7 @@ func testPermsStore_GrantPendingPermissions(db *sql.DB) func(t *testing.T) {
 								Perm:   authz.Read,
 							},
 						}, {
-							accounts: &ExternalAccounts{
+							accounts: &extsvc.ExternalAccounts{
 								ServiceType: "sourcegraph",
 								ServiceID:   "https://sourcegraph.com/",
 								AccountIDs:  []string{"bob"},
@@ -1261,7 +1262,7 @@ func testPermsStore_GrantPendingPermissions(db *sql.DB) func(t *testing.T) {
 					},
 					pendings: []pending{
 						{
-							accounts: &ExternalAccounts{
+							accounts: &extsvc.ExternalAccounts{
 								ServiceType: "sourcegraph",
 								ServiceID:   "https://sourcegraph.com/",
 								AccountIDs:  []string{"alice@example.com"},
@@ -1271,7 +1272,7 @@ func testPermsStore_GrantPendingPermissions(db *sql.DB) func(t *testing.T) {
 								Perm:   authz.Read,
 							},
 						}, {
-							accounts: &ExternalAccounts{
+							accounts: &extsvc.ExternalAccounts{
 								ServiceType: "sourcegraph",
 								ServiceID:   "https://sourcegraph.com/",
 								AccountIDs:  []string{"alice2@example.com"},
@@ -1434,7 +1435,7 @@ func testPermsStore_DeleteAllUserPendingPermissions(db *sql.DB) func(t *testing.
 
 		ctx := context.Background()
 
-		accounts := &ExternalAccounts{
+		accounts := &extsvc.ExternalAccounts{
 			ServiceType: "sourcegraph",
 			ServiceID:   "https://sourcegraph.com/",
 			AccountIDs:  []string{"alice", "bob"},
@@ -1508,7 +1509,7 @@ func testPermsStore_DatabaseDeadlocks(db *sql.DB) func(t *testing.T) {
 			}
 		}
 		setRepoPendingPermissions := func(ctx context.Context, t *testing.T) {
-			accounts := &ExternalAccounts{
+			accounts := &extsvc.ExternalAccounts{
 				ServiceType: "sourcegraph",
 				ServiceID:   "https://sourcegraph.com/",
 				AccountIDs:  []string{"alice"},
