+++ enterprise/cmd/repo-updater/authz/perms_syncer.go
@@ -317,7 +317,7 @@ func (s *PermsSyncer) syncRepoPerms(ctx context.Context, repoID api.RepoID) erro
 	}
 	defer txs.Done(&err)
 
-	accounts := &edb.ExternalAccounts{
+	accounts := &extsvc.ExternalAccounts{
 		ServiceType: "sourcegraph",
 		ServiceID:   "https://sourcegraph.com/",
 		AccountIDs:  pendingBindUsernames,
