+++ internal/extsvc/types.go
@@ -3,10 +3,12 @@ package extsvc
 import (
 	"encoding/json"
 	"time"
+
+	otlog "github.com/opentracing/opentracing-go/log"
 )
 
 // ExternalAccount represents a row in the `user_external_accounts` table. See the GraphQL API's
-// corresponding fields for documentation.
+// corresponding fields in "ExternalAccount" for documentation.
 type ExternalAccount struct {
 	ID                  int32
 	UserID              int32
@@ -18,7 +20,7 @@ type ExternalAccount struct {
 
 // ExternalAccountSpec specifies a user external account by its external identifier (i.e., by
 // the identifier provided by the account's owner service), instead of by our database's serial
-// ID. See the GraphQL API's corresponding fields for documentation.
+// ID. See the GraphQL API's corresponding fields in "ExternalAccount" for documentation.
 type ExternalAccountSpec struct {
 	ServiceType string
 	ServiceID   string
@@ -31,6 +33,24 @@ type ExternalAccountSpec struct {
 type ExternalAccountData struct {
 	AuthData    *json.RawMessage
 	AccountData *json.RawMessage
+}
+
+// ExternalAccounts contains a list of accounts that belong to the same external service.
+// All fields have a same meaning to ExternalAccountSpec. See GraphQL API's corresponding
+// fields in "ExternalAccount" for documentation.
+type ExternalAccounts struct {
+	ServiceType string
+	ServiceID   string
+	AccountIDs  []string
+}
+
+// TracingFields returns tracing fields for the opentracing log.
+func (s *ExternalAccounts) TracingFields() []otlog.Field {
+	return []otlog.Field{
+		otlog.String("ExternalAccounts.ServiceType", s.ServiceType),
+		otlog.String("ExternalAccounts.Perm", s.ServiceID),
+		otlog.Int("ExternalAccounts.AccountIDs.Count", len(s.AccountIDs)),
+	}
 }
 
 // ExternalAccountID is a descriptive type for the external identifier of an external account
